Warning hunting
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@5679 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/EDClient.cpp
@@ -746,7 +746,7 @@ void CEDClient::SendHello(BYTE nType)
 	CEDTag( ED2K_CT_SOFTWAREVERSION, nVersion ).Write( pPacket );
 
 	// 4 - Feature Versions. 
-	BYTE nExtendedRequests = min ( Settings.eDonkey.ExtendedRequest, (DWORD)ED2K_VERSION_EXTENDEDREQUEST );
+	BYTE nExtendedRequests = (BYTE)min ( Settings.eDonkey.ExtendedRequest, (DWORD)ED2K_VERSION_EXTENDEDREQUEST );
 	nVersion = ( ( ED2K_VERSION_AICH << 29) |			// AICH
 				 ( TRUE << 28) |						// Unicode
 				 ( ED2K_VERSION_UDP << 24) |			// UDP version
